Add steps directory
